requirements.txt
#Heroku
Django>=1.10.6
gunicorn>=19.7.0
psycopg2==2.6.2
dj-database-url==0.4.1
celery>=4.0.2
gevent>=1.2.1
whitenoise>=3.3.0
Pillow>=4.0.0
Markdown>=2.6.8
django-pagedown>=0.1.1
djangorestframework>=3.6.2
django-rest-swagger>=0.2.9
bleach>=2.0.0
django-redis-cache>=1.5.0
python-dotenv==0.6.1
click==6.7
boto>=2.46.1
django-queued-storage>=0.8.0
django-storages>=1.5.2
